semantic model improvements
